ZIntegralASimple input now includes figr option
The ZIntegralASimple function now includes a varargin variable so that the user can use the visualization option of the ZIntegralA function. This will require ZIntegralASimple description changes. This is the first step in introducing the varargin in the ZXX functions, which may eventually eliminate the need for ZXSimple functions all together.