Start of ZDerivativeFourier function development
The development of the ZDerivativeFourier function begins with this commit. It contains the ZDerivativeFourier function under development as well as the first two test scripts ZDerivativeFourierTestScript1 and ZDerivativeFourierTestScript2, used for the development of it. The ZDerivativeFourier function seemingly performs as intended but some testing, extensive description update, and major optimization are still required.