Major ZIntegralA input change – introduction of optional variables
The ZIntegralA was changed so only the xData, yData, xmin and xmax are required as input. The other three variables have been made optional and have been renamed. The description will thus have to be updated. This makes the ZIntegralASimple function obsolete and thus it was deleted. To test the changes of the ZIntegralA function being made, the ZIntegralTestScript1 test script was introduced again.