Beginning of major restructuring and utility of ZIntegralXXX functions
Several changes to the organization and the utility of ZIntegralXXX functions are being made that are planned to culminate in a single directory containing multiple functions and tutorial scripts, which can be used for numerical interpolation, differentiation and integration. In this commit, the GetPoints functions have been extracted from the ZIntgralA function and a subfunction (the SmatrixIpointsFunctionValue functon) of the not yet written general interpolation (/differentiation/integration ???) function as well as its first test script and the appropriate subfunction, the ConstructPolynomial function, have been written. They seemingly perform as intended, though significant optimization and description update is still required.