SmatrixIpointsDerivativeValue function development start
The SmatrixIpointsDerivativeValue development begins with this commit. It contains the SmatrixIpointsDerivativeValue function as well as the first test script for it, the SmatrixIpointsDerivativeValueTestScript1 script, and the auxiliary function, the ConstructDerivativePolynomial function. They seemingly perform as intended, though some optimization and significant description update is still required. To make some functions more general, the GetPolynomialCoefficients function was written and functions such as the SmatrixIpointsFunctionValue function and the SmatrixIpointsIntegralValue were extensively updated.