ZIntegralSpline revision start
The revision of the ZIntegralSpline function begins with this commit. The ZIntegralSpline_Revised function will eventually become the revised ZIntegralSpline function. The ZIntegralSpline_RevisedTestScript1 script is the first test script written for the task. The revised version already seemingly performs as intended but considerate work is still needed with regards to description and the actual code.