ZIntegralA revision start
The revision of the ZIntegralA function begins with this commit. The ZIntegralA_Revised function will eventually become the revised ZIntegralA function. The ZIntegralA function was divided into two functions, the ZIntegralA_Revised function and the ZIntegralALimits, and then broken down into a large number of subfunction and sub-subfunction and sub-sub-subfunctions, some of which will be in use in the revision of other functions in the package_integrate file. Some improvements and extensive description writing are still required for all of them.