Writing of function ZIndefiniteIntegralSpline in progress
With this commit, the ZIndefiniteIntegralSpline function and its subfunctions are in the process of being written for the first time. As is, the function seems to perform as intended, however, work is required on the parsing of input variables and documentation. Some minor updates of the names of the variables of some other functions were also made.