Major file movement for fit_ differentiate_integrate
With this commit, major renaming and moving of various files was executed. The fit, differentiation, and integration functions, and associated files were all moved to a single folder, the fit_differentiate_integrate folder, a folder for legacy files associated with the fit_differentiate_integrate folder was also created and appropriate files were moved there.