Major changes in distribution-propagation function package
With this commit, the ComposedVariable function was replaced by the FindOutputVariableAvgSigma function and the FindOutputVariableAvgCovarMat function, each of which includes calls to the two other newly written functions: the GetOutputVariablesMatrix function and the DrawOutputVariableDistribution function. Also, the DrawSimpleLinearRegressionGraph function was updated to account for the covariance of the slope and the intercept of the linear function. A new test script, the TestScript_Ver4 script, was written to test these new functions.