FindSimpleLinearRegressionCoefficients option added
With this commit, the input parser and the option to find the variance of the yData vector or the covariance matrix of the slope and the intercept was added to the FindSimpleLinearRegressionCoefficients function. Also, a faster version of the FindSimpleLinearRegressionCoefficients function, the FindSimpleLinearRegressionCoefficientsFAST function, was written without an input parser or optional input or output parameters.