GetRndValues update
With this commit, the GetRndValues function was updated so that vectors can be used as means in individual elements of the input VariablesDistributionInfo cell.
Also, two scripts, both of which combine the functions of the distribution-propagation function package and the functions of the fit-differentiate-integrate function package, were written: the FitPropagationScript script and the DerivativePropagationScript script.