File/Function rearrangement
With this commit, the distribution-propagation file was rearranged: a file for the tutorial and the subfunctions as well as a README.md file were created. Also, a new Physics file file was created: the other-tools file, which is to contain some of the useful functions and scripts that have accumulated over time.